requirements.txt
PySimpleGUI
ptoaster
aiofiles==0.4.0
anyio==1.2.3
arrow==0.15.5
asks==2.3.6
async-generator==1.10
atomicwrites==1.3.0
attrs==19.3.0
blinker==1.4
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
colorama==0.4.3
h11==0.9.0
h2==3.1.1
hpack==3.0.0
hstspreload==2020.1.15
Hypercorn==0.9.0
hyperframe==5.2.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
more-itertools==8.1.0
multidict==4.7.4
outcome==1.0.1
packaging==20.0
pluggy==0.13.1
priority==1.3.0
py==1.8.1
pycparser==2.19
pyparsing==2.4.6
pytest==5.3.3
python-dateutil==2.8.1
Quart==0.10.0
requests==2.22.0
rfc3986==1.3.2
six==1.14.0
sniffio==1.1.0
sortedcontainers==2.1.0
SQLAlchemy==1.3.12
toml==0.10.0
tqdm==4.41.1
trio==0.13.0
typing-extensions==3.7.4.1
urllib3==1.25.7
wcwidth==0.1.8
wsproto==0.15.0

pip~=19.0.3
wheel~=0.33.6
cryptography~=2.8
lxml~=4.4.2
setuptools~=40.8.0
boltons~=20.0.0
pandas~=0.25.3
datefinder~=0.7.0
parsel~=1.5.2
Scrapy~=1.8.0